+++ lib/phobos/handler.rb
@@ -6,7 +6,7 @@ module Phobos
       base.extend(ClassMethods)
     end
 
-    def before_consume(payload)
+    def before_consume(payload, metadata)
       payload
     end
 
@@ -14,14 +14,14 @@ module Phobos
       raise NotImplementedError
     end
 
+    def around_consume(payload, metadata)
+      yield
+    end
+
     module ClassMethods
       def start(kafka_client); end
 
       def stop; end
-
-      def around_consume(_payload, _metadata)
-        yield
-      end
     end
   end
 end