+++ spec/lib/phobos/actions/process_message_spec.rb
@@ -7,6 +7,23 @@ RSpec.describe Phobos::Actions::ProcessMessage do
     include Phobos::Handler
   end
 
+  # For backwards compatibility
+  class TestHandler2 < Phobos::EchoHandler
+    include Phobos::Handler
+
+    def self.around_consume(payload, metadata)
+      yield
+    end
+  end
+
+  class TestHandler3 < Phobos::EchoHandler
+    include Phobos::Handler
+
+    def before_consume(payload)
+      payload
+    end
+  end
+
   let(:payload) { 'message-1234' }
   let(:topic) { 'test-topic' }
   let(:message) do
@@ -33,12 +50,55 @@ RSpec.describe Phobos::Actions::ProcessMessage do
     )
   end
 
+  before do
+    allow(subject).to receive(:sleep) # Prevent sleeping in tests
+  end
+
   it 'processes the message by calling around consume, before consume and consume of the handler' do
-    expect(TestHandler).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
-    expect_any_instance_of(TestHandler).to receive(:before_consume).with(payload).once.and_call_original
+    expect_any_instance_of(TestHandler).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
+    expect_any_instance_of(TestHandler).to receive(:before_consume).with(payload, subject.metadata).once.and_call_original
     expect_any_instance_of(TestHandler).to receive(:consume).with(payload, subject.metadata).once.and_call_original
 
     subject.execute
+  end
+
+  context '.around_consumed defined' do
+    let(:listener) do
+      Phobos::Listener.new(
+        handler: TestHandler2,
+        group_id: 'test-group',
+        topic: topic
+      )
+    end
+
+    it 'supports and prefers around_consume if defined as a class method' do
+      expect(TestHandler2).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
+      expect(Phobos).to receive(:deprecate).once
+      expect_any_instance_of(TestHandler2).to receive(:before_consume).with(payload, subject.metadata).once.and_call_original
+      expect_any_instance_of(TestHandler2).to receive(:consume).with(payload, subject.metadata).once.and_call_original
+
+      subject.execute
+    end
+  end
+
+  context '#before_consume defined with 1 argument' do
+    let(:listener) do
+      Phobos::Listener.new(
+        handler: TestHandler3,
+        group_id: 'test-group',
+        topic: topic
+      )
+    end
+
+    it 'supports the method and logs a deprecation message' do
+      expect(Phobos).to receive(:deprecate).once
+      expect_any_instance_of(TestHandler3).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
+      expect_any_instance_of(TestHandler3).to receive(:before_consume).with(payload).once.and_call_original
+      expect_any_instance_of(TestHandler3).to receive(:consume).with(payload, subject.metadata).once.and_call_original
+
+      subject.execute
+    end
+
   end
 
   context 'with encoding' do
@@ -112,6 +172,54 @@ RSpec.describe Phobos::Actions::ProcessMessage do
         end.to raise_error(Phobos::AbortError)
 
         expect(subject.metadata[:retry_count]).to eq(0)
+      end
+    end
+  end
+
+  describe '#snooze' do
+    context 'when interval is 0' do
+      let(:interval) { 0 }
+
+      it 'attempts to send a heartbeat' do
+        expect(listener).to receive(:send_heartbeat_if_necessary).once
+        subject.snooze(interval)
+      end
+
+      it 'does not sleep' do
+        expect(subject).to receive(:sleep).never
+        subject.snooze(interval)
+      end
+    end
+
+    context 'when interval is less than MAX_SLEEP_INTERVAL' do
+      let(:interval) { described_class::MAX_SLEEP_INTERVAL.to_f / 2 }
+
+      it 'attempts to send heartbeats twice' do
+        expect(listener).to receive(:send_heartbeat_if_necessary).twice
+        subject.snooze(interval)
+      end
+
+      it 'sleeps once at the specified interval' do
+        expect(subject).to receive(:sleep).with(interval).once
+        subject.snooze(interval)
+      end
+    end
+
+    context 'when interval is MAX_SLEEP_INTERVAL + 1' do
+      let(:interval) { described_class::MAX_SLEEP_INTERVAL + 1 }
+
+      it 'attempts to send heartbeats 3 times' do
+        expect(listener)
+          .to receive(:send_heartbeat_if_necessary)
+          .exactly(3).times
+        subject.snooze(interval)
+      end
+
+      it 'sleeps once for MAX_SLEEP_INTERVAL and once for 1 second' do
+        expect(subject).to receive(:sleep).with(described_class::MAX_SLEEP_INTERVAL).ordered.once
+        expect(subject).to receive(:sleep).with(1).ordered.once
+        expect(subject).to receive(:sleep).ordered.never
+        subject.snooze(interval)
       end
     end
   end